completed appData singleton up until assets lair

+++ app/src/main/java/com/example/jessexx/panic/Models/Asset.java
@@ -16,7 +16,7 @@ public class Asset {
 
     private int pictureID;
 
-    private float value;
+    private double value;
 
     private boolean tangible;
 
@@ -27,7 +27,7 @@ public class Asset {
 
     public Asset(){}
 
-    public Asset(String name, String description, int pictureID, float value, CATEGORY category) {
+    public Asset(String name, String description, int pictureID, double value, CATEGORY category) {
         this.name = name;
         this.description = description;
         this.pictureID = pictureID;
@@ -37,7 +37,7 @@ public class Asset {
     }
 
 
-    public Asset(String name, int pictureID, float value, CATEGORY category) {
+    public Asset(String name, int pictureID, double value, CATEGORY category) {
         this.id = UUID.randomUUID();
         this.name = name;
         this.pictureID = pictureID;
@@ -73,7 +73,7 @@ public class Asset {
         this.pictureID = pictureID;
     }
 
-    public float getValue() {
+    public double getValue() {
         return value;
     }
