+++ app/src/main/java/com/example/jessexx/panic/Models/Bank.java
@@ -5,7 +5,7 @@ package com.example.jessexx.panic.Models;
  */
 public class Bank extends Asset {
 
-    public Bank(String name, String description, int PictureId, float value, CATEGORY category) {
+    public Bank(String name, String description, int PictureId, double value, CATEGORY category) {
         super(name, description, PictureId, value, category);
     }
